typecheck decorator
