+++ pyproject.toml
@@ -125,6 +125,9 @@ strict = true
 addopts = [
   "--import-mode=importlib", # allow using test files with same name
 ]
+filterwarnings = [
+  "error",
+]
 strict = true
 testpaths = [ "tests" ]
